Extract array operator helpers
